# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Customer sign-up API. Endpoint specs live in the\nhandler annotations, not in README.md.",
        "title": "golang-gin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log in with an email or a social account",
                "parameters": [
                    {
                        "description": "mediaName decides which fields are required",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "wrong credentials, or social token rejected",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "403": {
                        "description": "customer is not active",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "social account verified but not registered",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign up with an email or a social account",
                "parameters": [
                    {
                        "description": "mediaName decides which fields are required",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.CustomerResponse"
                        }
                    },
                    "400": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "social token rejected, or mediaId does not match the verified profile",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "409": {
                        "description": "email or social account already registered",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Liveness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.CustomerResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.LoginRequest": {
            "type": "object",
            "required": [
                "mediaName"
            ],
            "properties": {
                "email": {
                    "description": "Email login only.",
                    "type": "string",
                    "maxLength": 255
                },
                "mediaId": {
                    "description": "Social login only.",
                    "type": "string",
                    "maxLength": 255
                },
                "mediaName": {
                    "enum": [
                        "email",
                        "kakao",
                        "naver"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.MediaName"
                        }
                    ]
                },
                "mediaToken": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 128
                }
            }
        },
        "domain.LoginResponse": {
            "type": "object",
            "properties": {
                "customer": {
                    "$ref": "#/definitions/domain.CustomerResponse"
                },
                "expires_at": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "domain.MediaName": {
            "type": "string",
            "enum": [
                "email",
                "kakao",
                "naver"
            ],
            "x-enum-varnames": [
                "MediaEmail",
                "MediaKakao",
                "MediaNaver"
            ]
        },
        "domain.SignUpRequest": {
            "type": "object",
            "required": [
                "email",
                "mediaName"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "mediaId": {
                    "type": "string",
                    "maxLength": 255
                },
                "mediaName": {
                    "enum": [
                        "email",
                        "kakao",
                        "naver"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.MediaName"
                        }
                    ]
                },
                "mediaToken": {
                    "type": "string"
                },
                "name": {
                    "description": "Social sign-up only. The name comes from the verified profile, so it is\nrequired only for the email flow.",
                    "type": "string",
                    "maxLength": 64
                },
                "password": {
                    "description": "Email sign-up only.",
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8
                }
            }
        },
        "response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}